// myApp/src/app/models/horario.ts
export interface Horario {
  id?:  any;
  nomecompleto: string;
  entrada_Segunda:  string;
  intervalo_Segunda:  string;
  saida_Segunda:  string;
  entrada_Terca:  string;
  intervalo_Terca:  string;
  saida_Terca:  string;
  entrada_Quarta:  string;
  intervalo_Quarta:  string;
  saida_Quarta:  string;
  entrada_Quinta:  string;
  intervalo_Quinta:  string;
  saida_Quinta:  string;
  entrada_Sexta:  string;
  intervalo_Sexta:  string;
  saida_Sexta:  string;
  entrada_Sabado:  string;
  intervalo_Sabado:  string;
  saida_Sabado:  string;
  observacao:  string;
  turno:  string;
  setor: string;
  empresa: string;
  horariototalsemanal: string;
  horariototaldiario: string;
}